requirements.txt
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.1.0
openai>=1.0
weaviate-client>=3.24
fastapi>=0.104
uvicorn>=0.24
pydantic>=2.0
yfinance>=0.2.28
python-dotenv>=1.0
streamlit>=1.28
httpx>=0.25
tiktoken>=0.5
pyyaml>=6.0
